OSStatus is an SInt32; use the correct formatting.
BUG=177744
TEST=compiles in 64 bit mode

Review URL: https://codereview.chromium.org/12747006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@188068 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: abbb4130f02afb5d305ecf9c0e27dbb7a41529a2

+++ remoting/host/installer/mac/uninstaller/remoting_uninstaller_app.mm
@@ -51,7 +51,8 @@
       message = @"Chrome Remote Desktop Host uninstall authorization denied.";
     } else {
       [NSException raise:@"AuthorizationCopyRights Failure"
-          format:@"Error during AuthorizationCopyRights status=%ld", status];
+                  format:@"Error during AuthorizationCopyRights status=%d",
+                             static_cast<int>(status)];
     }
     if (message != NULL) {
       NSLog(@"Uninstall %s: %@", success ? "succeeded" : "failed", message);
@@ -93,7 +94,7 @@ int main(int argc, char* argv[])
       OSStatus status = [uninstaller remotingUninstall];
 
       NSLog(@"Chrome Remote Desktop Host uninstall complete.");
-      NSLog(@"Status = %ld", status);
+      NSLog(@"Status = %d", static_cast<int>(status));
       return status != errAuthorizationSuccess;
     }
   } else {
